refactor: split drive item actions
